chunks_app/requirements.txt
asgiref==3.8.1
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
Django==5.1.7
djangorestframework==3.15.2
filelock==3.17.0
fsspec==2025.3.0
huggingface-hub==0.29.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==76.0.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
